pyproject.toml
[tool.poetry]
name = "hw-from-les-24"
version = "0.1.0"
description = ""
authors = ["mikhapol_linux <mikhapol@bk.ru>"]
readme = "README.md"
packages = [{include = "hw_from_les_24"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
psycopg2-binary = "^2.9.7"
python-dotenv = "^1.0.0"
pillow = "^10.0.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
